+++ src/operations/misc/norm.cpp
@@ -17,10 +17,10 @@
 namespace hicma
 {
 
-double norm(const Matrix& A) { return norm_omm(A); }
+float norm(const Matrix& A) { return norm_omm(A); }
 
-define_method(double, norm_omm, (const Dense& A)) {
-  double l2 = 0;
+define_method(float, norm_omm, (const Dense& A)) {
+  float l2 = 0;
   timing::start("Norm(Dense)");
   for (int64_t i=0; i<A.dim[0]; i++) {
     for (int64_t j=0; j<A.dim[1]; j++) {
@@ -31,10 +31,10 @@ define_method(double, norm_omm, (const Dense& A)) {
   return l2;
 }
 
-define_method(double, norm_omm, (const LowRank& A)) { return norm(Dense(A)); }
+define_method(float, norm_omm, (const LowRank& A)) { return norm(Dense(A)); }
 
-define_method(double, norm_omm, (const Hierarchical& A)) {
-  double l2 = 0;
+define_method(float, norm_omm, (const Hierarchical& A)) {
+  float l2 = 0;
   for (int64_t i=0; i<A.dim[0]; i++) {
     for (int64_t j=0; j<A.dim[1]; j++) {
       l2 += norm(A(i, j));
@@ -43,7 +43,7 @@ define_method(double, norm_omm, (const Hierarchical& A)) {
   return l2;
 }
 
-define_method(double, norm_omm, (const Matrix& A)) {
+define_method(float, norm_omm, (const Matrix& A)) {
   omm_error_handler("norm", {A}, __FILE__, __LINE__);
   std::abort();
 }
